+++ first/src/main/java/graduation/first/oauth/controller/AuthController.java
@@ -9,8 +9,8 @@ import graduation.first.oauth.entity.UserPrincipal;
 import graduation.first.oauth.token.AuthToken;
 import graduation.first.oauth.token.AuthTokenProvider;
 import graduation.first.common.utils.HeaderUtil;
-import graduation.first.user.UserRefreshToken;
-import graduation.first.user.UserRefreshTokenRepository;
+import graduation.first.user.domain.UserRefreshToken;
+import graduation.first.user.repository.UserRefreshTokenRepository;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
