+++ first/src/main/java/graduation/first/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -9,8 +9,8 @@ import graduation.first.oauth.info.OAuth2UserInfoFactory;
 import graduation.first.oauth.repository.OAuth2AuthorizationRequestBasedOnCookieRepository;
 import graduation.first.oauth.token.AuthToken;
 import graduation.first.oauth.token.AuthTokenProvider;
-import graduation.first.user.UserRefreshToken;
-import graduation.first.user.UserRefreshTokenRepository;
+import graduation.first.user.domain.UserRefreshToken;
+import graduation.first.user.repository.UserRefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
