+++ first/src/main/java/graduation/first/oauth/service/PrincipalOAuth2UserService.java
@@ -7,8 +7,8 @@ import graduation.first.oauth.exception.OAuthErrorCode;
 import graduation.first.oauth.exception.OAuthException;
 import graduation.first.oauth.info.OAuth2UserInfo;
 import graduation.first.oauth.info.OAuth2UserInfoFactory;
-import graduation.first.user.User;
-import graduation.first.user.UserRepository;
+import graduation.first.user.domain.User;
+import graduation.first.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.core.AuthenticationException;
