+++ first/src/main/java/graduation/first/post/PostController.java
@@ -1,8 +1,10 @@
 package graduation.first.post;
 
 import graduation.first.oauth.entity.UserAdapter;
-import graduation.first.oauth.entity.UserPrincipal;
-import graduation.first.user.User;
+import graduation.first.post.dto.PostResponseDto;
+import graduation.first.post.dto.PostResponseVO;
+import graduation.first.post.dto.PostSaveRequestDto;
+import graduation.first.post.dto.PostUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
