+++ first/src/main/java/graduation/first/post/PostRepository.java
@@ -1,12 +1,10 @@
 package graduation.first.post;
 
 import graduation.first.category.Category;
-import graduation.first.user.User;
+import graduation.first.user.domain.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
