+++ first/src/main/java/graduation/first/post/PostService.java
@@ -5,7 +5,15 @@ import graduation.first.category.CategoryErrorCode;
 import graduation.first.category.CategoryException;
 import graduation.first.category.CategoryRepository;
 import graduation.first.oauth.entity.UserAdapter;
-import graduation.first.user.*;
+import graduation.first.post.dto.PostResponseDto;
+import graduation.first.post.dto.PostResponseVO;
+import graduation.first.post.dto.PostSaveRequestDto;
+import graduation.first.post.dto.PostUpdateRequestDto;
+import graduation.first.post.exception.PostErrorCode;
+import graduation.first.post.exception.PostException;
+import graduation.first.user.domain.User;
+import graduation.first.user.exception.UserErrorCode;
+import graduation.first.user.exception.UserException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
