+++ first/src/main/java/graduation/first/post/dto/PostResponseVO.java
@@ -1,11 +1,9 @@
-package graduation.first.post;
+package graduation.first.post.dto;
 
+import graduation.first.post.Post;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.data.domain.Page;
-
-import java.util.ArrayList;
-import java.util.List;
 
 @Getter
 @AllArgsConstructor
