+++ first/src/main/java/graduation/first/post/exception/PostException.java
@@ -1,4 +1,4 @@
-package graduation.first.post;
+package graduation.first.post.exception;
 
 import lombok.Getter;
