+++ first/src/main/java/graduation/first/user/UserService.java
@@ -2,13 +2,13 @@ package graduation.first.user;
 
 import graduation.first.post.Post;
 import graduation.first.post.PostRepository;
+import graduation.first.user.domain.User;
+import graduation.first.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
