+++ first/src/main/java/graduation/first/user/repository/UserRepository.java
@@ -1,6 +1,7 @@
-package graduation.first.user;
+package graduation.first.user.repository;
 
 
+import graduation.first.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
